pyproject.toml
[tool.poetry]
name = "neopay_api"
version = "0.1.0"
description = ""
authors = ["Гильфанов Руслан <rgilfanov@fix.ru>"]

[tool.poetry.dependencies]
python = "3.9.1"
SQLAlchemy = "1.4.0b1"
fastapi = "^0.63.0"
alembic = "^1.5.2"
psycopg2-binary = "2.8.6"
telegram-log = "^1.0.12"
sentry-sdk = "^0.19.5"
raven = "^6.10.0"
pytz = "^2020.5"
uvicorn = "^0.13.3"
SQLAlchemy-Utils = "^0.36.8"
python-jose = "^3.2.0"
passlib = "^1.7.4"
bcrypt = "^3.2.0"
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
